Refining controllers for Site Notices
- Refining create action

- Refining update action

- Refining dismiss_notice action

+++ app/controllers/admin/site_notices_controller.rb
@@ -8,55 +8,71 @@ class Admin::SiteNoticesController < ApplicationController
   end
 
   def create
-    site_notice = SiteNotice.new
-    site_notice.notice_html = params[:notice_html]
-    site_notice.creator_id = current_user.id
-    site_notice.save!
-    
-    #storing all roles that should see this notice
-    roles = params[:role]
-    roles.each do |role_id|
-      site_notice_roles = SiteNoticeRole.new
-      site_notice_roles.notice_id = site_notice.id
-      site_notice_roles.role_id = role_id  
-      site_notice_roles.save!
+    if params[:notice_html] =~ /\S+/
+      if !params[:role].nil?
+        site_notice = Admin::SiteNotice.new
+        site_notice.notice_html = params[:notice_html]
+        site_notice.created_by = current_user.id
+        site_notice.save!
+        
+        #storing all roles that should see this notice
+        
+        roles = params[:role]
+        roles.each do |role_id|
+          site_notice_roles = Admin::SiteNoticeRole.new
+          site_notice_roles.notice_id = site_notice.id
+          site_notice_roles.role_id = role_id
+          site_notice_roles.save!
+        end
+      else
+        flash[:error] = "Select atleast one role"
+      end       
+    else
+      flash[:error] = "Text cannot be blank"
     end
     redirect_to admin_site_notices_path
   end
   
   def index
-    @all_notices = SiteNotice.all
+    @all_notices = Admin::SiteNotice.all
   end
   
   def edit
-    @notice = SiteNotice.find(params[:id])
-    @notice_roles = SiteNoticeRole.find_all_by_notice_id(params[:id])
+    @notice = Admin::SiteNotice.find(params[:id])
+    @notice_roles = Admin::SiteNoticeRole.find_all_by_notice_id(params[:id])
     @notice_role_ids = @notice_roles.map{|notice_role| notice_role.role_id}
   end
   
   def update
-      #Storing new html for notice
-      site_notice = SiteNotice.find(params[:id])
-      site_notice.notice_html= params[:notice_html]
-      site_notice.updated_by = current_user.id
-      site_notice.save!
-      
-      notice_roles = SiteNoticeRole.find_all_by_notice_id(params[:id])
-      notice_roles.each do |notice_role|
-        notice_role.destroy
+    #Storing new html for notice
+    if params[:notice_html] =~ /\S+/
+      if !params[:role].nil?
+        site_notice = Admin::SiteNotice.find(params[:id])
+        site_notice.notice_html= params[:notice_html]
+        site_notice.updated_by = current_user.id
+        site_notice.save!
+          
+        notice_roles = Admin::SiteNoticeRole.find_all_by_notice_id(params[:id])
+        notice_roles.each do |notice_role|
+          notice_role.destroy
+        end
+          
+        #Storing new roles
+        roles = params[:role]
+        roles.each do |role_id|
+          site_notice_roles = Admin::SiteNoticeRole.new
+          site_notice_roles.notice_id = params[:id]
+          site_notice_roles.role_id = role_id 
+          site_notice_roles.save!
+        end
+      else
+        flash[:error] = "Select atleast one role"
       end
-      
-      #Storing new roles
-      roles = params[:role]
-      roles.each do |role_id|
-        site_notice_roles = SiteNoticeRole.new
-        site_notice_roles.notice_id = params[:id]
-        site_notice_roles.role_id = role_id 
-        site_notice_roles.save! 
-      end
+    else
+      flash[:error] = "Text cannot be blank"
+    end  
        
-      redirect_to admin_site_notices_path
-      
+    redirect_to admin_site_notices_path
   end
    
     
@@ -68,7 +84,12 @@ class Admin::SiteNoticesController < ApplicationController
         notice_role.destroy
       end
       
-      SiteNotice.find(params[:id]).destroy
+    notice_users = Admin::SiteNoticeUser.find_all_by_notice_id(params[:id])
+    notice_users.each do |notice_user|
+      notice_user.destroy
+    end
+        
+    Admin::SiteNotice.find(params[:id]).destroy
       
       if request.xhr?
         render :update do |page|
@@ -79,7 +100,54 @@ class Admin::SiteNoticesController < ApplicationController
       
       #redirect_to :action => 'index';
       #redirect_to admin_site_notices_path
-   end
+  end
   
     
+
+  def toggle_notice_display
+    user_collapsed_notice = Admin::NoticeUserDisplayStatus.find_or_create_by_user_id(current_user.id)
+    status_to_be_set = (user_collapsed_notice.collapsed_status.nil? || user_collapsed_notice.collapsed_status == false)? true : false
+    dateTime = Time.new
+    collapsed_timestamp = dateTime.to_time
+    user_collapsed_notice.last_collapsed_at_time = collapsed_timestamp
+    user_collapsed_notice.collapsed_status = status_to_be_set
+    user_collapsed_notice.save!
+    if request.xhr?
+      render :update do |page|
+        if status_to_be_set
+           page << "$('user_notice_container_div').hide()";
+           page << "$('oHideShowLink').update('Show Notices')";
+        else
+          page << "$('user_notice_container_div').show();"
+          page << "$('oHideShowLink').update('Hide Notices')";
+        end
+      end
+      return 
+    end
+  end
+  
+  def dismiss_notice
+    notice = Admin::SiteNotice.find(params[:id])
+    user_notice = Admin::SiteNoticeUser.new
+    user_notice.user_id = current_user.id
+    user_notice.notice_id = params[:id]
+    user_notice.notice_dismissed = 1
+    user_notice.save!
+    if request.xhr?
+      render :update do |page|
+        page << "$('#{dom_id_for(notice)}').remove();"
+        page << "notice_table = document.getElementById('all_notice_to_render')
+                  all_notices = notice_table.getElementsByTagName('tr')
+                  if(all_notices.length == 1)
+                  {
+                    $('oHideShowLink').remove();
+                    $('user_notice_container_div').remove();
+                    $('notice_container').innerHTML = 'No Notice';
+                  }
+                "
+      end
+      return
+    end    
+  end  
+
 end
